+++ src/com/landawn/abacus/DataSet.java
@@ -4789,7 +4789,7 @@ public interface DataSet {
 
     /**
      * 
-     * @return key are column name, value is column - an immutable list.
+     * @return key are column name, value is column - an immutable list, backed by the column in this {@code DataSet}.
      */
     Map<String, List<Object>> toColumnMap();